update all feature authen now build frontend

+++ src/main/java/com/Nguyen/blogplatform/repository/RoleRepository.java
@@ -10,7 +10,8 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByName(ERole name);
     boolean existsByName(ERole name);
 }
+
